fix(agent): split probeSnapshotter into Linux/non-Linux files (syscall.Mount is Linux-only)
